Changed disarming event
Changed how the disarming event is handled, currently it would just destroy your item. That's not disarming, that's a headache for server owners having to replace the items. 
Original Code: 
				}
				if(RNG(15)==14) {
					player.getInventory().setItemInMainHand(new ItemStack(Material.AIR));
					player.sendMessage("You have been disarmed...");
				}
Updated Code: Checks if items in hand, ignores if there's not; Drops item that is in hand and replaces it with air.
				}
                if(RNG(15)==14) {
                    ItemStack heldItem = player.getInventory().getItemInMainHand();
                    if (heldItem != null && heldItem.getType() != Material.AIR) {
                        player.getWorld().dropItemNaturally(player.getLocation(), heldItem);
                        player.getInventory().setItemInMainHand(new ItemStack(Material.AIR));
                        player.sendMessage("You have been disarmed...");
                    }