Handles bossbars better
Made it so whenever a boss is killed, it gets rid of the bossbar instead of needing to log out for the boss bars to disappear. 

Added Dr.Strange & Piggy to have Boss Bars. 

Fixed the boss bar health reading, so bosses read the config.yml for how much health is on the boss bar instead of showing that it has a very small amount of health left despite being full health.